Refactor Model to separate files
