+++ src/main/kotlin/astminer/common/ModelTreeSplitting.kt
@@ -1,0 +1,6 @@
+package astminer.common
+
+
+interface TreeSplitter<T : Node> {
+    fun split(root: T): Collection<T>
+}
